+++ tests/test_before_hooks.py
@@ -43,7 +43,7 @@ def parse_body(req, resp, resource, params):
 
     length = req.content_length or 0
     if length != 0:
-        params['doc'] = json.load(io.TextIOWrapper(req.stream, 'utf-8'))
+        params['doc'] = json.load(io.TextIOWrapper(req.bounded_stream, 'utf-8'))
 
 
 def bunnies(req, resp, resource, params):